Renamed admin to board
